added comments

+++ app.js
@@ -67,24 +67,37 @@ const upload = multer({
 
 
 // TODO: Backend checking: need to check again if all fields are valid
+// FIXME: if want a button that goes back to home, need to change this else ok
 app.post('/signup', upload.single('profilepic'), async (req, res) => {
     const { firstname, lastname, email, number, password, confirmpassword } = req.body;
+    const formData = { firstname, lastname, email, number };
     const profilepic = req.file;
 
     if (!profilepic) {
         return res.status(400).send('Profile picture is required');
     }
 
+    //TODO: checkers for valid name, email and number
+    if (password !== confirmpassword) {
+        //FIXME: [DOES NOT WORK] return to signup page, restore filled out fields, and indicate an error message
+        return res.redirect('/signup');
+    }
+
     try{
         const fileType = await import('file-type');
         const fileBuffer = fs.readFileSync(profilepic.path);
         const fileTypeResult = await fileType.fileTypeFromBuffer(fileBuffer);
+
         if (!fileTypeResult || !fileTypeResult.mime.startsWith('image/')) {
             fs.unlinkSync(profilepic.path); // Delete the invalid file
-            return res.status(400).send('Invalid img');
+            //FIXME: [DOES NOT WORK] return to signup page, restore filled out fields, and indicate an error message
+            return res.redirect('/signup')
         }
-        else{
-            console.log('Valid img')
+        else{ //Signup success
+            //TODO: Put to db
+
+            //Back to login
+            return res.redirect('/login');
         }
     }catch(err){
         console.error('Error loading file-type module or processing image: ', err);
